# internal/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package api

// ErrorResponse Standard error response
type ErrorResponse struct {
	Error string `json:"error"`
}

// FizzBuzzResponse Generated FizzBuzz sequence
type FizzBuzzResponse struct {
	Result []string `json:"result"`
}

// StatsResponse Most frequent FizzBuzz request statistics
type StatsResponse struct {
	Hits  int    `json:"hits"`
	Int1  int    `json:"int1"`
	Int2  int    `json:"int2"`
	Limit int    `json:"limit"`
	Str1  string `json:"str1"`
	Str2  string `json:"str2"`
}

// GenerateFizzBuzzParams defines parameters for GenerateFizzBuzz.
type GenerateFizzBuzzParams struct {
	// Int1 First divisor
	Int1 int `form:"int1" json:"int1"`

	// Int2 Second divisor
	Int2 int `form:"int2" json:"int2"`

	// Limit Upper bound of the sequence
	Limit int `form:"limit" json:"limit"`

	// Str1 Replacement string for multiples of int1
	Str1 string `form:"str1" json:"str1"`

	// Str2 Replacement string for multiples of int2
	Str2 string `form:"str2" json:"str2"`
}